extend reconcile transactions for stripe cards
